requirements.txt
setuptools~=49.6.0
matplotlib~=3.4.3
numpy~=1.19.2
Pillow~=8.0.1
tqdm~=4.52.0
Pebble~=4.6.1
ghostscript~=0.6
scikit-image~=0.17.2
streamlit~=0.71.0
PyMuPDF~=1.18.4
pytesseract~=0.3.6
pandas~=1.1.2
spacy~=2.3.4
psutil~=5.7.2
plotly~=4.14.3
kaleido~=0.1.0
Pweave~=0.30.3
dill~=0.3.3
scikit-learn~=0.23.2
pdfminer-six==20201018
opencv-python~=4.5.1.48
imagehash~=4.2.0
func-timeout~=4.3.5
jpype1~=1.2.0
mmh3~=3.0.0